refactor(pds): remove direct upload functionality and restructure related code
